requirements.txt
black==24.10.0
click==8.1.7
kafka-python==2.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
py4j==0.10.9.7
pyspark==3.5.3
